Add files via upload
We are Team ABA. Our name comes from our last names: Amoroso, Balagosa, and Atienza. We want our logo to show a girl to represent femininity. We also prefer a minimalist design for our logo. The simplicity of the design reflects our approach—clean, straightforward, and impactful. We believe that less is more, and we want our logo to be elegant yet powerful, embodying the strength and grace of femininity.